melhorando a performance do Banner
