+++ leadsoft-frontend/src/app/page.tsx
@@ -5,7 +5,7 @@ import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import Inscricao from "@/components/Inscricao";
 import dynamic from "next/dynamic";
-import Banner from "@/components/Banner";
+import Banner from "@/components/Banner/Banner";
 import HeaderClient from "@/components/Header/HeaderClient";
 
 const Galeria = dynamic(() => import("@/components/Galeria"), {
